requirements.txt
feedparser>=6.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
schedule>=1.2.0
